requirements.txt
python-dotenv
aiogram
stripe
asyncio
fastapi
pyngrok
uvicorn